backend/requirements.txt
# FastAPI ve Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Veri İşleme
pandas==2.1.4
numpy==1.26.3

# Deep Learning ve Sentetik Veri
torch==2.1.2
sdv==1.9.0

# Yardımcı Kütüphaneler
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Güvenlik
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Test
pytest==7.4.4
httpx==0.26.0

# PII Detection ve Sentetik Veri Üretimi (Faz 2)
faker==22.6.0
spacy==3.7.2
presidio-analyzer==2.2.351
presidio-anonymizer==2.2.351

# Differential Privacy (Faz 2.2)
diffprivlib==0.6.4

# Visualization and Reporting (Faz 3)
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
scikit-learn==1.4.0

# Asynchronous Task Processing (Faz 4.1)
celery==5.3.6
redis==5.0.1
flower==2.0.1
